+++ src/atividade8/Retangulo.java
@@ -3,23 +3,23 @@ package atividade8;
 import javax.swing.JOptionPane;
 
 public class Retangulo {
-	
-	private double lado1;
-	
-	private double lado2;
-	
-	private double area;
-	
-	private double perimetro;
-	
-	
+
+	private float lado1;
+
+	private float lado2;
+
+	private float area;
+
+	private float perimetro;
+
+
 	public Retangulo() {}
-	
-	public Retangulo(double lado1, double lado2) {
+
+	public Retangulo(float lado1, float lado2) {
 		this.lado1 = lado1;
 		this.lado2 = lado2;
 	}
-	
+
 	public void calcularArea() {
 		this.area = (this.lado1 * this.lado2);
 		JOptionPane.showMessageDialog(null, "A área do retangulo é: "+this.area);
@@ -29,37 +29,38 @@ public class Retangulo {
 		JOptionPane.showMessageDialog(null, "O perimetro do retangulo é: "+this.perimetro);
 	}
 
-	public double getLado1() {
+	public float getLado1() {
 		return lado1;
 	}
 
-	public void setLado1(double lado1) {
+	public void setLado1(float lado1) {
 		this.lado1 = lado1;
 	}
 
-	public double getLado2() {
+	public float getLado2() {
 		return lado2;
 	}
 
-	public void setLado2(double lado2) {
+	public void setLado2(float lado2) {
 		this.lado2 = lado2;
 	}
 
-	public double getArea() {
+	public float getArea() {
 		return area;
 	}
 
-	public void setArea(double area) {
+	public void setArea(float area) {
 		this.area = area;
 	}
 
-	public double getPerimetro() {
+	public float getPerimetro() {
 		return perimetro;
 	}
 
-	public void setPerimetro(double perimetro) {
+	public void setPerimetro(float perimetro) {
 		this.perimetro = perimetro;
 	}
 
-	
+
+
 }
